Route renderer clipboard writes through the main process
The session permission handler allows only geolocation and speaker
selection, so navigator.clipboard.writeText rejects with NotAllowedError
inside the app window. Copy buttons appeared to do nothing.

An IPC bridge already existed but only two of six call sites used it.
Collect the fallback chain in one helper and use it everywhere, so a
failed copy reports the value instead of failing silently.
